requirements.txt
python-dotenv
pandas
requests
streamlit
Pillow
tabulate
plotly
matplotlib
sqlite3